pyproject.toml
[project]
name = "web-inspector-mcp"
version = "0.1.0"
description = "Web Inspector MCP Server"
dependencies = [
    "mcp",
    "pydoll-python"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cov=web_inspector_mcp --cov-report=term-missing --cov-report=xml"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["web_inspector_mcp"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
